Bundling of pure ruby json lib as "pson"
Bundeling and renaming the pure ruby json library to addresses a
number of cross version serliaization bugs (#2615, et al).

This patch adds a subset of the files from the json_pure gem to
lib/puppet/external/pson (renamed to avoid conflicts with rails) so
that we will always have a known-good erialization format available.

The pure ruby json gem as distibuted defers to the compiled version
if it is installed.  This is problematic in some circumstances so the
files that have been brought over have been modified to always and
only use the bundled version.

It's a large patch, so here's a breakdown of the change categories:

The majority of the lines are only marginally interesting:

* The json lib itself (in lib/puppet/external/pson) make up the bulk
  of the lines.
* Renaming of json to pson make up the second largest group.

Somewhat more interesting are the following, which can be located by
searching the diffs for the indicated strings:

* Adjusting tests to reflect the changes
* Changing the encoding/decoding behavior so that nested structures
  (e.g. resources) don't serialize as escaped strings.  This should
  make it much easier to process the results with external tools, if
  needed.  Search for "to_pson" and "to_pson_data_hash"
* Cleaning up the envelope/metadata
    * Now provides a document_type (as opposed to a ruby class name) by
      using a symple registration scheme instead of constant lookup
      (search for "document_type")
    * Added an api_version (search for "api_version")
    * Added a hash for document metadata (search for "metadata")
* Removing the yaml monkeypatch and instead disabling yaml serialization
  on ruby 1.8.1 in favor of pson (search for "yaml")
* Cleaning up the json/rails feature interaction (they're now totally
  independent) (search for "feature")